requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
appopener==1.7
beautifulsoup4==4.12.3
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
comtypes==1.4.5
confection==0.1.5
cymem==2.0.8
dateparser==1.2.0
idna==3.7
Jinja2==3.1.4
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
numpy==1.26.4
packaging==24.1
pefile==2023.2.7
preshed==3.0.9
psutil==6.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyqtspinner==2.0.0
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
regex==2024.7.24
requests==2.32.3
rich==13.7.1
schedule==1.2.2
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sounddevice==0.4.7
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
srsly==2.4.8
thinc==8.2.5
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
uuid==1.30
wasabi==1.1.3
weasel==0.4.1
word2number==1.1
wrapt==1.16.0
